requirements.txt
python-multipart~=0.0.9
jinja2~=3.1.2
uvicorn~=0.27.0.post1
fastapi~=0.109.0
pydantic~=2.6.0
python-docx~=1.1.0
requests~=2.28.1
openai~=1.9.0
starlette~=0.35.1
pypdf~=4.1.0